Added support for the piecewise element in top-level expressesions.
The existing top-level Equation abstraction is replaced in the code with a MathStatement abstraction, which covers top-level Equation, Piecewise, and
Inequality entities. Piecewise is supported efficiently for single solutions where the structure of the model is not affected by the contents of the
piecewise. Support for where the piecewise is a 'reset rule' for a differential equation is anticipated, but not yet implemented.