Fixed tracker item 2070: Based on Andrew Miller's suggestion in comment 2 of that tracker item (call computeRates before calling computeAlgebraic).
Matlab style is to do vectorised arithmetic, which is why computeAlgebraic is called after solve has returned.  So, call computeRates directly before calling computeAlgebraic.  This was needed anyway, because the "intermediate" algebraics were not being saved anyway, due to the way Matlab ODE solvers work.  Note that this was all prompted by having a working version of Saucerman model in CellML, and the approach taken in the original Saucerman Matlab was to compute all algebraics (even ones not needed for computing rates) via the DAE/MassMatrix approach, which is perhaps the way Matlab expects intermediate algebraics to be handled if their values are required after ODE solve is complete.