Fix reference counting race conditions (causing problems with SProS, but potentially also some CIS based programs). Also update the uncertain parameter support to use new long form element names.
This change makes a class for reference counts that provides mutex-wrapped reference counting.

A side-effect of this change was to allow explicit reference count initialisation to be replaced with implicit reference counting. As a result, nearly every file in the code base was changed to remove the explicit reference count initialisations that are no longer needed.

This fix also addresses a few minor exception safety issues in SRuS.