Fix a problem that arose from the tracker item 3297 fix - track threads we use Java on, and ensure we tell the JVM using DetachCurrentThread
This created a need for a cleanup of Utilities.hxx and ThreadWrapper.hxx (now removed and merged into Utilities.hxx) to give a general
facility for supporting destructors to be called when a thread is destroyed. This facility is now also used by ThreadLocalStorage to
call the destructor for thread local structures on Windows.

Consequential to this change, the dead code in dom_gdome, which depended on ThreadWrapper.hxx has been removed, and Utilities.cpp has
been populated with some code that didn't make sense in the header, which is now linked into CellML.

Some dead code has also been removed from Utilities.hxx in the process.